// src/api/panapplications/entities/pan.entity.ts
export enum applicationTypes {
  NEW = "NEW",
  CORRECTION = "CORRECTION",
}

export enum title {
  MR = "Mr.",
  MRS = "Mrs.",
  MISS = "Miss",
}
export enum parantType {
  FATHER = "FATHER",
  MOTHER = "MOTHER",
}
export enum flowStatus {
  STATUS_CHANGED = "STATUS_CHANGED",
  ASSIGNEE_CHANGED = "ASSIGNEE_CHANGED",
}

enum paymentCategory {
  BASIC_SERVICE = "BASIC_SERVICE",
}
export enum paymentCategories {
  BASIC_SERVICE = paymentCategory.BASIC_SERVICE,
  CAT_DSA = categoryCode.DSA,
  CAT_PCO = categoryCode.PCO,
  ADDITIONAL_CAT_1 = categoryCode.cat_2,
  ADDITIONAL_CAT_2 = categoryCode.cat_3,
}
export enum status {
  BLANK = "PAYMENT_PENDING",
  PENDING = "PENDING",
  IN_PROGRESS = "IN_PROGRESS",
  VERIFY = "VERIFY",
  REJECT = "REJECT",
  GENERATE = "GENERATE",
  DONE = "DONE",
  CANCELLED = "CANCELLED",
}

export enum formCategory {
  BLANK = "",
  INDIVIDUAL = "INDIVIDUAL",
  ASSOCIATION_OF_PERSONS = "ASSOCIATION OF PERSONS",
  BODY_OF_INDIVIDUALS = "BODY OF INDIVIDUALS",
  COMPANY = "COMPANY",
  TRUST = "TRUST",
  LIMITED_LIABILITY_PARTNERSHIP = "LIMITED LIABILITY PARTNERSHIP",
  FIRM = "FIRM",
  GOVERNMENT = "GOVERNMENT",
  HINDU_UNDIVIDED_FAMILY = "HINDU UNDIVIDED FAMILY",
  ARTIFICIAL_JURIDICAL_PERSON = "ARTIFICIAL JURIDICAL PERSON",
  LOCAL_AUTHORITY = "LOCAL AUTHORITY",
}

export enum urlFields {
  panCardFront = "Pan card front photo url",
  passportPhotoUrl = "Passport photo url",
  signaturePhotoUrl = "Signature photo url",
  panFormFrontPhotoUrl = "Pan form front photo url",
  panFormBackPhotoUrl = "Pan form back photo url",
  adhaarFrontPhotoUrl = "Adhaar front photo url",
  adhaarBackPhotoUrl = "Adhaar back photo url",
}